GinTub: Database sprocs
Added a script to delete and recreate the GinTub database. Updated the
script to create the GinTub database tables a bit. Added a script to
create the FIRST round of dev sprocs. I'll definitely need more--like,
having the capacity to remove entries. Which will be a little
complicated, due to dependencies, but not unmanageable. Basically, I can
just loop through all dependent objects with a cursor or whatever, since
I don't need to worry about performance in a dev environment like this.