GinTub: Builder
Fulfilled my promise from yesterday: Controls now implement an interface
that allows someone else to tell them to register and unregister GinTub
events. This allows me to turn off the LocationGet event for the
Image_DisplayLocation when I'm on the Areas and am creating a new
RoomState.

I was going to remove the Image_DisplayLocation class and simply merge
it with the Grid_RoomModify, but for now it's fine the way it is.