GinTub: Builder
Some refinement to Verbs; specifically, they can now be modified with
the Grid_VerbData element. Although, this probably won't see much use,
since I want this to be a mouse-only game. Also, working on adding
Results. Had to change things up a bit, but I'm happier with the
results: Now Results have JSONData, and a ResultType, and the ResultType
is linked to a table in the [dev] schema that lists what JSON fields
should be present.